finished customization

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ use App\Like;
 use App\Post;
 use App\Tag;
 use Auth;
+use Gate;
 use Illuminate\Http\Request;
 
 class PostController extends Controller
@@ -18,6 +19,9 @@ class PostController extends Controller
 
     public function getAdminIndex()
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $posts = Post::orderBy('title', 'asc')->get();
         return view('admin.index', ['posts' => $posts]);
     }
@@ -38,12 +42,18 @@ class PostController extends Controller
 
     public function getAdminCreate()
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $tags = Tag::all();
         return view('admin.create', ['tags' => $tags]);
     }
 
     public function getAdminEdit($id)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $post = Post::find($id);
         $tags = Tag::all();
         return view('admin.edit', ['post' => $post, 'postId' => $id, 'tags' => $tags]);
@@ -71,11 +81,17 @@ class PostController extends Controller
 
     public function postAdminUpdate(Request $request)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10'
         ]);
         $post = Post::find($request->input('id'));
+        if (Gate::denies('manipulate-post', $post)) {
+            return redirect()->back();
+        }
         $post->title = $request->input('title');
         $post->content = $request->input('content');
         $post->save();
@@ -87,7 +103,13 @@ class PostController extends Controller
 
     public function getAdminDelete($id)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $post = Post::find($id);
+        if (Gate::denies('manipulate-post', $post)) {
+            return redirect()->back();
+        }
         $post->likes()->delete();
         $post->tags()->detach();
         $post->delete();
